+++ main.go
@@ -1,7 +1,10 @@
 package main
 
 import (
+	dockerassets "github.com/faradey/madock/v3/docker"
+	scriptassets "github.com/faradey/madock/v3/scripts"
 	"github.com/faradey/madock/v3/src/app"
+	"github.com/faradey/madock/v3/src/helper/embedded"
 	"github.com/faradey/madock/v3/src/version"
 
 	// Register all built-in controllers
@@ -9,5 +12,7 @@ import (
 )
 
 func main() {
+	embedded.SetDockerFS(dockerassets.FS)
+	embedded.SetScriptsFS(scriptassets.FS)
 	app.Run(version.Version)
 }