+++ scripts/embed.go
@@ -1,0 +1,6 @@
+package scriptassets
+
+import "embed"
+
+//go:embed all:php
+var FS embed.FS